requirements.txt
tensorflow-cpu==2.15.0
numpy
pillow
flask
flask-cors
gunicorn